requirements.txt
africastalking==1.2.4
asgiref==3.5.2
attrs==21.4.0
backports.zoneinfo==0.2.1
black==22.3.0
boto3==1.24.17
botocore==1.27.17
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
contextlib2==21.6.0
coverage==6.4.1
Django==4.0.5
django-appconf==1.0.5
django-imagekit==4.1.0
django-phonenumber-field==6.3.0
django-storages==1.12.3
djangorestframework==3.13.1
idna==3.3
iniconfig==1.1.1
isort==5.10.1
jmespath==1.0.1
mypy==0.961
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.50
pilkit==2.0
Pillow==9.1.1
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
schema==0.7.5
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
